Simplifying the SwapiResolver
Resolvers should be responsible for api calls forwarding to services.
Their logic should be as simple as possible. They can be treated as the
system boundary. The services should implement bussiness logic.
A proper interface was added to GraphqlService and part of the logic
from SwapiResolver was moved to GraphqlService.